+++ examples/full-screen/clock.py
@@ -1,33 +1,23 @@
 import datetime
 
-from quo.console import Console
 from quo import container
-from quo.keys import Bind
+from quo.keys import bind
 from quo.text import Text
-from quo.layout import Layout
-from quo.widget import Label
 from quo.layout import Window, FormattedTextControl as FC
+from quo.widget import Label
 
-now = datetime.datetime.now()
-
-bind = Bind()
-
-c_time = "%s:%s:%s" %  (now.hour, now.minute, now.second)
-
-def get_prompt():
-    "Tokens to be shown before the prompt."
+def get_time():
     now = datetime.datetime.now()
     return [
-            ("bg:green fg:black", "%s:%s:%s"  % (now.hour, now.minute, now.second)),
-            ("bg:cornsilk fg:maroon", "Enter something: ")
+            ("bg:green fg:black", "%s:%s:%s"  % (now.hour, now.minute, now.second))
             ]
-content = Window(FC(get_prompt))
 
-lay = Layout(content)
+content = Window(FC(get_time), align="center")
+
 
 @bind.add("ctrl-c")
 def _(event):
     event.app.exit()
 
-Console(layout=lay, bind=bind, refresh_interval=0.8).run()
-#ontainer(content, bind=True, full_screen=True)
+#Console(layout=lay, bind=bind, refresh_interval=0.8).run()
+container(content,  full_screen=True)