+++ examples/getchar.py
@@ -1,0 +1,3 @@
+from quo.getchar import getchar
+
+getchar()