+++ examples/pause.py
@@ -1,0 +1,4 @@
+from quo import pause
+
+
+pause(err=True)