+++ examples/print-text/shadow.py
@@ -1,0 +1,6 @@
+from quo import container
+
+
+from quo.widget import Shadow, Label
+
+container(Shadow(Label("dke")), bind=False)