+++ examples/print-text/styles-text-tuples1.py
@@ -1,0 +1,10 @@
+from quo import print
+from quo.text import FormattedText
+
+text = FormattedText([
+    ('fg:red', 'Hello'),
+    ('', ' '),
+    ('fg:purple italic', 'World')
+    ])
+
+print(text, fmt=True)