+++ examples/prompts/input-validation.py
@@ -2,15 +2,18 @@
 """
 Simple example of input validation.
 """
-import quo
 
-session = quo.Prompt()
+from quo.prompt import Prompt
+from quo.types import Validator
+
+
+session = Prompt()
 
 def is_valid_email(text):
     return "@" in text
 
 
-validator = quo.types.Validator.from_callable(
+validator = Validator.from_callable(
     is_valid_email,
     error_message="Not a valid e-mail address (Does not contain an @.",
     move_cursor_to_end=True,
@@ -20,13 +23,13 @@ validator = quo.types.Validator.from_callable(
 def main():
     # Validate when pressing ENTER.
     text = session.prompt(
-        "Enter e-mail address: ", validator=validator, validate_while_typing=False
+        "Enter e-mail address: ", type=validator, validate_while_typing=False
     )
     print("You said: %s" % text)
 
     # While typing
     text = session.prompt(
-        "Enter e-mail address: ", validator=validator, validate_while_typing=True
+        "Enter e-mail address: ", type=validator, validate_while_typing=True
     )
     print("You said: %s" % text)